# swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "App Store Connect Adapters",
    "description": "HTTP API for querying an App Store Connect adapters host.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:48189"
    }
  ],
  "paths": {
...
    "/api/app-store-connect/v1.0/events/{adapterId}/by-time-range": {
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Reads historical event messages for the specified time range.",
        "operationId": "Events_ReadEventMessagesForTimeRange",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter ID.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The search filter.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadEventMessagesForTimeRangeRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "Successful responses contain a collection of EventMessage objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventMessage"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/app-store-connect/v1.0/events/{adapterId}/by-cursor": {
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Reads historical event messages starting at the specified cursor position.",
        "operationId": "Events_ReadEventMessagesByCursor",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter ID.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The search filter.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadEventMessagesUsingCursorRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "Successful responses contain a collection of EventMessageWithCursorPosition objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventMessageWithCursorPosition"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/app-store-connect/v1.0/events/{adapterId}/write": {
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Writes event messages to an adapter. Up to MaxEventMessagesPerWriteRequest messages can be written in a single request.",
        "description": "Up to MaxEventMessagesPerWriteRequest values can be written to the adapter in a single request. Subsequent values will be ignored. No corresponding WriteEventMessageResult object will be returned for these items.",
        "operationId": "Events_WriteEventMessages",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter ID to write to.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The event messages to write.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteEventMessagesRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "Successful responses contain a collection of WriteEventMessageResult objects (one per sample written).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WriteEventMessageResult"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdapterRequest": {
        "type": "object",
        "description": "Base class that adapter request objects can inherit from.",
        "x-abstract": true,
        "additionalProperties": false,
        "properties": {
          "properties": {
            "type": "object",
            "description": "Additional request properties. These can be used to provide bespoke query parameters \nsupported by the adapter.",
            "nullable": true,
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "AdapterProperty": {
        "type": "object",
        "description": "Describes a custom property associated with an adapter, tag, tag value, event message, etc.",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The property name.",
            "minLength": 1
          },
          "value": {
            "description": "The property value.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/Variant"
              }
            ]
          },
          "description": {
            "type": "string",
            "description": "The property description.",
            "nullable": true
          }
        }
      },
      "Variant": {
        "type": "object",
        "description": "Describes a variant value.",
        "additionalProperties": false,
        "properties": {
          "value": {
            "description": "The value.",
            "nullable": true
          },
          "type": {
            "description": "The variant type.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/VariantType"
              }
            ]
          }
        }
      },
      "VariantType": {
        "type": "string",
        "description": "Describes the type of a variant value.",
        "x-enumNames": [
          "Unknown",
          "Null",
          "Object",
          "Boolean",
          "SByte",
          "Byte",
          "Int16",
          "UInt16",
          "Int32",
          "UInt32",
          "Int64",
          "UInt64",
          "Float",
          "Double",
          "String",
          "DateTime",
          "TimeSpan",
          "Url"
        ],
        "enum": [
          "Unknown",
          "Null",
          "Object",
          "Boolean",
          "SByte",
          "Byte",
          "Int16",
          "UInt16",
          "Int32",
          "UInt32",
          "Int64",
          "UInt64",
          "Float",
          "Double",
          "String",
          "DateTime",
          "TimeSpan",
          "Url"
        ]
      },
...
      "EventMessage": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EventMessageBase"
          },
          {
            "type": "object",
            "description": "Describes a message generated by e.g. an alarms &amp; events system.",
            "additionalProperties": false
          }
        ]
      },
      "EventMessageBase": {
        "type": "object",
        "description": "Base class describing a message generated by e.g. an alarms &amp; events system.",
        "x-abstract": true,
        "additionalProperties": false,
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "A unique identifier for the event.",
            "minLength": 1
          },
          "topic": {
            "type": "string",
            "description": "The topic for the event (when the event is emitted by an adapter that supports \ntopic-based event messages).",
            "nullable": true
          },
          "utcEventTime": {
            "type": "string",
            "description": "The UTC timestamp of the event.",
            "format": "date-time"
          },
          "priority": {
            "description": "The priority associated with the event.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/EventPriority"
              }
            ]
          },
          "category": {
            "type": "string",
            "description": "The event category.",
            "nullable": true
          },
          "message": {
            "type": "string",
            "description": "The event message.",
            "nullable": true
          },
          "properties": {
            "type": "array",
            "description": "Additional event properties.",
            "items": {
              "$ref": "#/components/schemas/AdapterProperty"
            }
          }
        }
      },
      "EventPriority": {
        "type": "string",
        "description": "Describes the priority associated with an event message.",
        "x-enumNames": [
          "Unknown",
          "Low",
          "Medium",
          "High",
          "Critical"
        ],
        "enum": [
          "Unknown",
          "Low",
          "Medium",
          "High",
          "Critical"
        ]
      },
      "ReadEventMessagesForTimeRangeRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ReadHistoricalEventMessagesRequest"
          },
          {
            "type": "object",
            "description": "Describes a request to retrieve historical event messages using a time range.",
            "additionalProperties": false,
            "required": [
              "utcStartTime",
              "utcEndTime"
            ],
            "properties": {
              "topics": {
                "type": "array",
                "description": "The topics to read messages for. This property will be ignored if the adapter does not \nsupport a topic-based event model.",
                "maxItems": 100,
                "nullable": true,
                "items": {
                  "type": "string"
                }
              },
              "utcStartTime": {
                "type": "string",
                "description": "The UTC start time for the request.",
                "format": "date-time",
                "minLength": 1
              },
              "utcEndTime": {
                "type": "string",
                "description": "The UTC end time for the request.",
                "format": "date-time",
                "minLength": 1
              },
              "pageSize": {
                "type": "integer",
                "description": "The page size for the query.",
                "format": "int32",
                "default": 10,
                "maximum": 2147483647.0,
                "minimum": 1.0
              },
              "page": {
                "type": "integer",
                "description": "The page number for the query.",
                "format": "int32",
                "default": 1,
                "maximum": 2147483647.0,
                "minimum": 1.0
              }
            }
          }
        ]
      },
      "ReadHistoricalEventMessagesRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AdapterRequest"
          },
          {
            "type": "object",
            "description": "Base class for adapter requests related to reading historical event messages.",
            "x-abstract": true,
            "additionalProperties": false,
            "properties": {
              "direction": {
                "description": "The event read direction. When Backwards is specified, \nthe resulting events will be returned in descending order of time.",
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/EventReadDirection"
                  }
                ]
              }
            }
          }
        ]
      },
      "EventReadDirection": {
        "type": "string",
        "description": "Describes the read direction for a historical event message read operation.",
        "x-enumNames": [
          "Forwards",
          "Backwards"
        ],
        "enum": [
          "Forwards",
          "Backwards"
        ]
      },
      "EventMessageWithCursorPosition": {
        "allOf": [
          {
            "$ref": "#/components/schemas/EventMessageBase"
          },
          {
            "type": "object",
            "description": "Describes a message generated by e.g. an alarms &amp; events system that also contains a cursor position.",
            "additionalProperties": false,
            "required": [
              "cursorPosition"
            ],
            "properties": {
              "cursorPosition": {
                "type": "string",
                "description": "The cursor position for the event message.",
                "minLength": 1
              }
            }
          }
        ]
      },
      "ReadEventMessagesUsingCursorRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ReadHistoricalEventMessagesRequest"
          },
          {
            "type": "object",
            "description": "Describes a request to retrieve historical event messages using a cursor to specify the query start position.",
            "additionalProperties": false,
            "properties": {
              "topic": {
                "type": "string",
                "description": "The topic to read messages for. This property will be ignored if the adapter does not \nsupport a topic-based event model.",
                "nullable": true
              },
              "cursorPosition": {
                "type": "string",
                "description": "The cursor position to start the query at.",
                "nullable": true
              },
              "pageSize": {
                "type": "integer",
                "description": "The page size for the query.",
                "format": "int32",
                "default": 10,
                "maximum": 2147483647.0,
                "minimum": 1.0
              }
            }
          }
        ]
      },
      "WriteEventMessageResult": {
        "allOf": [
          {
            "$ref": "#/components/schemas/WriteOperationResult"
          },
          {
            "type": "object",
            "description": "Describes the result of an event message write operation.",
            "additionalProperties": false,
            "properties": {
              "correlationId": {
                "type": "string",
                "description": "The optional correlation ID for the operation.",
                "nullable": true
              }
            }
          }
        ]
      },
      "WriteOperationResult": {
        "type": "object",
        "description": "Base class for result objects returned from write operations.",
        "x-abstract": true,
        "additionalProperties": false,
        "properties": {
          "status": {
            "description": "Indicates if the write was successful.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/WriteStatus"
              }
            ]
          },
          "notes": {
            "type": "string",
            "description": "Notes associated with the write.",
            "nullable": true
          },
          "properties": {
            "type": "array",
            "description": "Additional properties related to the write.",
            "items": {
              "$ref": "#/components/schemas/AdapterProperty"
            }
          }
        }
      },
      "WriteStatus": {
        "type": "string",
        "description": "Describes the status of a write operation.",
        "x-enumNames": [
          "Unknown",
          "Success",
          "Fail",
          "Pending"
        ],
        "enum": [
          "Unknown",
          "Success",
          "Fail",
          "Pending"
        ]
      },
      "WriteEventMessagesRequest": {
        "type": "object",
        "description": "Describes a request to write event messages to an adapter.",
        "additionalProperties": false,
        "required": [
          "events"
        ],
        "properties": {
          "events": {
            "type": "array",
            "description": "The event messages to write.",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/WriteEventMessageItem"
            }
          }
        }
      },
      "WriteEventMessageItem": {
        "type": "object",
        "description": "Describes an event message being written to an adapter.",
        "additionalProperties": false,
        "required": [
          "eventMessage"
        ],
        "properties": {
          "correlationId": {
            "type": "string",
            "description": "The optional correlation ID for the operation.",
            "nullable": true
          },
          "eventMessage": {
            "description": "The event message to write.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/EventMessage"
              }
            ]
          }
        }
      },
...
    }
  }
}